requirements.txt
protobuf==6.30.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.43.0
google-generativeai
